eden: add PathMap type
Summary:
The intention is to use this for the PathTrie that is in a following diff.

The motivation is to avoid making a N copies of a string when traversing a path
of depth N in a PathTrie.  The PathMap is backed by a vector and knows how to
compare the Piece variant of our path types.

I have not performed any rigorous comparisons, but based on reading around
other flat_map vs map comparisons, the expectation is that this will use less
memory (and avoid extraneous allocations during lookups) than a regular map,
with most random access being fairly similar.  Inserts need to move the memory
around to make room and are thus slower than a regular map.  Sorted inserts
perform the best for PathMap as that translates into repeated back-insertions.

My intuition is that this is the right trade off to make for managing the
overlay data structures; the size of these materialized dirs is likely going to
be small compared to the full fbsource tree.

Reviewed By: simpkins

Differential Revision: D3470377

fbshipit-source-id: 6f4f3abe33818da6981d8ab821ebe004a78d382c