sqlite: keep nonnormal files in the separate table
Summary:
The low cardinality of the status column makes it basically
non-indexable (try to do a binsearch in a sorted array with
1m numbers from the range {1..5} and you will see why this approach
doesn't work).

There is a feature we could use in sqlite 3.9 called partial indexes - so
we could index only the files that have non-normal statuses.

The sqlite 3.6 that we are using on centos6 unfortunately doesn't support it.
IMO the cost of shipping and maintaining our separate version of sqlite is
too high so let's simulate it by keeping those entries in separate table.

Test Plan:
 * tests are passing
 * on fbsource/fbcode "hg st" went down from ~800ms to ~600ms.

Reviewers: #mercurial, ttung, durham

Subscribers: mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3559244