eden: add debug logging of exception reasons for failed fuse requests
Summary:
It's useful in some debugging situations to understand the context for something like
`folly::throwSystemErrorExplicit(ENOENT, "useful context")`.

This is conceptually as simple as adding some `VLOG` statements in the error
handler for a fuse request, but I thought that it would be worth while to clean
up the macro that `clang-format` was doing a poor job of formatting.  In turn,
I noticed that we can simplify the cancellation handling a bit while we're in
here, since we're only ever dealing with `Future<Unit>` in here now.

This reduces some boilerplate in Dispatcher.cpp.

Reviewed By: simpkins

Differential Revision: D3630958

fbshipit-source-id: 8d9dd50797f421ab9794b87f10d4c02c96e6fea8