update DirList to accept StringPiece names
Summary:
Update DirList::add() to accept entry names as a folly::StringPiece instead of
just a null-terminated string.  This makes it possible to use PathComponent
objects without having to make a copy of the name just to ensure that it is
null terminated.

The libfuse APIs unfortunately only accept null terminated strings, so we now
simply manually populate the fuse_dirent structs in our buffer rather than
using the libfuse helper methods.

Reviewed By: wez

Differential Revision: D3762623

fbshipit-source-id: d4132b354912e0e003090bddcad0ce912f4ed401