ctree: move file iterator to a separate function
Summary:
This moves the file iterator to a separate function so that it may be reused in
other areas. It also further separates the iterator from the python specific
logic.

Two notable changes to the logic:
1. It now returns the node and flag as well as the path.

2. Instead of copying iter->path straight into a python string for returning,
the caller provides a path, node, and flag character array as arguments and the
result is copied into that array. This results in a little more string copying
during iteration, but since the character array is stack allocated it the actual
performance impact is neglible.

Test Plan:
Ran the perf test, verified fulliter perf didn't change (2.2s before
and after).  Also used pdb to verify the results were valid and expected paths.

Reviewers: #fastmanifest

Differential Revision: https://phabricator.intern.facebook.com/D3760071