sshaskpass: do not import mercurial packages when running standalone
Summary:
The sshaskpass code is both an hg extension and a standalone python script.
In the latter case, it does not need to import mercurial packages (and
doing that may cause `ImportError` if `mercurial` is not in python path).

Test Plan:
Manually make the change in a Mac Mini device and check `chg pull` works
with ssh password authentication.

Reviewers: #mercurial, ttung, durham

Reviewed By: durham

Subscribers: mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3799653

Signature: t1:3799653:1472675282:9f7e9f4f29281db239f33cf396057ea12ff21773