logginghelper: rename reporootlog to logginghelper
Summary:
We are going to log many different pieces of information in this extensions.
Therefore the `reporootlog` is not an appropriate name anymore.
Change it to `logginghelper`.

Test Plan:
Note that repo root is still logged instead of repo name. It will be changed in subsequent diffs
{P56637657}



Reviewers: rmcelroy, quark, ttung

Reviewed By: ttung

Subscribers: mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3840985

Signature: t1:3840985:1473410718:62d83f008f2973b7ed5266b0a5ebf67fe54b48c4