fbhistedit: use `ui.system` instead of `util.system`
Summary:
`chg` has some special handling on `ui.system` - if it is simple, it will
be forwarded to the client (the `chg` process) and the new process will be
a child of `chg`, which will inherit the controlling TTY correctly.

Otherwise, it will be a child of a server-side worker process, which does
not have a controlling TTY. That's the case if `util.system` is used.

Test Plan:
Insert `x sleep 1000` in the histedit plan and confirm the `sleep`
process runs as a child of the `chg` process:

```
41876 13176 pts/14   Ss   03:17   0:00  \_ zsh
 4196   732 pts/14   S+   03:20   0:00  |   \_ chg histedit
 7572   664 pts/14   S+   03:20   0:00  |       \_ sleep 1000
```

Reviewers: #mercurial, ttung, rmcelroy, simpkins

Reviewed By: simpkins

Subscribers: mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3846243

Tasks: 13295639

Signature: t1:3846243:1473480297:59aaf34cb3a52200321122dc7c054e2454cd3bd3